misc: reorganize net and parsers directory
