+++ src/net/dash/index.js
@@ -16,7 +16,7 @@
 
 import { Observable } from "rxjs/Observable";
 
-import { resolveURL } from "../../../utils/url";
+import { resolveURL } from "../../utils/url";
 
 import {
   parseSidx,
