+++ src/net/dash/manifest/__tests__/index.test.js
