+++ src/net/dash/manifest/filters.js
