+++ src/net/dash/utils.js
