+++ src/net/directfile/index.js
@@ -18,7 +18,6 @@ import { Observable } from "rxjs/Observable";
 
 const manifestPipeline = {
   loader({ url }) {
-    debugger;
     return Observable.of({ url });
   },
   parser({ url }) {
