+++ src/net/smooth/utils.js
