+++ src/parsers/isobmff.js
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import assert from "../../utils/assert";
+import assert from "../utils/assert";
 import {
   itobe4, be8toi, be4toi, be2toi, be3toi,
   hexToBytes, strToBytes, concat,
-} from "../../utils/bytes";
+} from "../utils/bytes";
 
 /**
  * Find the right atom (box) in an isobmff file from its hexa-encoded name.
